ci: use env instead of secret variable
It seems like the CI doesn't allow any PR to access the secrets store.
Since the Role Arn is not really a secret and the role itself is scoped
down to access the test key and nothing else, there is limited risk in
having it available as an env variable, but should hopefully allow for
testing PRs, although it does make setting up testing in a fork slightly
more challenging.

Signed-off-by: Petre Eftime <epetre@amazon.com>