Downgrade KMS SDK to avoid conflicts with TSS dependencies
In the latest version aws-sdk-kms and tss-esapi has a common conflicting
transitive dependency. Each of the libraries has different version requirements
for that dependency and they don't intersect.

To make build successful we align it and downgrade aws-sdk-kms to the version
where conflict disappears.

Also aws-config dependency is used only for testing and is moved to dev dependencies.

Bumping up the successfully buildable package version to publish it to crates.io

Signed-off-by: Eugene Koira <eugkoira@amazon.com>