Create Tokio runtime on the fly
Currently `KmsKey`s must be used from Tokio runtime. This commit adds a
check for the current handle, and in case methods like
`new_from_public_key`, `sign`, or `verify` are not called from a Tokio
thread, a Thread will be created and the code will be executed from that
context.

Signed-off-by: Mark Kirichenko <mkirich@amazon.de>